+++ src/main/java/top/girlkisser/lazuli/api/codec/LazuliExtraCodecs.java
@@ -5,6 +5,8 @@ import com.mojang.serialization.codecs.RecordCodecBuilder;
 import io.netty.buffer.ByteBuf;
 import net.minecraft.network.codec.ByteBufCodecs;
 import net.minecraft.network.codec.StreamCodec;
+import net.minecraft.world.phys.AABB;
+import net.minecraft.world.phys.Vec3;
 import org.jetbrains.annotations.ApiStatus;
 import org.joml.Vector2f;
 
@@ -30,4 +32,18 @@ public interface LazuliExtraCodecs
 	 * Stream codec for {@link #VECTOR2F}.
 	 */
 	StreamCodec<ByteBuf, Vector2f> VECTOR2F_STREAM_CODEC = ByteBufCodecs.fromCodec(VECTOR2F);
+
+	/**
+	 * A codec for an {@link AABB}. Uses {@link AABB#getMinPosition()} and {@link AABB#getMaxPosition()}
+	 * for serialization.
+	 */
+	Codec<AABB> AABB = RecordCodecBuilder.create(it -> it.group(
+		Vec3.CODEC.fieldOf("min").forGetter(net.minecraft.world.phys.AABB::getMinPosition),
+		Vec3.CODEC.fieldOf("max").forGetter(net.minecraft.world.phys.AABB::getMaxPosition)
+	).apply(it, AABB::new));
+
+	/**
+	 * Stream codec for {@link #AABB}.
+	 */
+	StreamCodec<ByteBuf, AABB> AABB_STREAM_CODEC = ByteBufCodecs.fromCodec(AABB);
 }